refactor: split route_optimizer.py into focused modules
Extract haversine distance + coordinate validation into
backend/agents/geo_utils.py, and move ICS/Maps URL builders
from route_export.py into backend/agents/export/{ics,maps_url}.py.

route_optimizer.py now contains only clustering and itinerary building.
Backward-compatible re-exports kept in route_optimizer and route_export
so all existing import sites and tests continue to work without changes.

Closes #143
