+++ backend/tests/unit/test_route_optimizer.py
@@ -61,42 +61,60 @@ def test_validate_rejects_missing_lat() -> None:
     rows: list[dict[str, object]] = [{"id": "a", "longitude": 139.0}]
     valid, invalid = validate_coordinates(rows)
     assert len(valid) == 0, f"Expected no valid rows for missing lat, got {len(valid)}"
-    assert len(invalid) == 1, f"Expected 1 invalid row for missing lat, got {len(invalid)}"
+    assert len(invalid) == 1, (
+        f"Expected 1 invalid row for missing lat, got {len(invalid)}"
+    )
 
 
 def test_validate_rejects_missing_lng() -> None:
     rows: list[dict[str, object]] = [{"id": "a", "latitude": 35.0}]
     valid, invalid = validate_coordinates(rows)
     assert len(valid) == 0, f"Expected no valid rows for missing lng, got {len(valid)}"
-    assert len(invalid) == 1, f"Expected 1 invalid row for missing lng, got {len(invalid)}"
+    assert len(invalid) == 1, (
+        f"Expected 1 invalid row for missing lng, got {len(invalid)}"
+    )
 
 
 def test_validate_rejects_out_of_range_lat() -> None:
     rows: list[dict[str, object]] = [{"id": "a", "latitude": 100.0, "longitude": 139.0}]
     valid, invalid = validate_coordinates(rows)
-    assert len(valid) == 0, f"Expected no valid rows for out-of-range lat, got {len(valid)}"
-    assert len(invalid) == 1, f"Expected 1 invalid row for out-of-range lat, got {len(invalid)}"
+    assert len(valid) == 0, (
+        f"Expected no valid rows for out-of-range lat, got {len(valid)}"
+    )
+    assert len(invalid) == 1, (
+        f"Expected 1 invalid row for out-of-range lat, got {len(invalid)}"
+    )
 
 
 def test_validate_rejects_out_of_range_lng() -> None:
     rows: list[dict[str, object]] = [{"id": "a", "latitude": 35.0, "longitude": 200.0}]
     valid, invalid = validate_coordinates(rows)
-    assert len(valid) == 0, f"Expected no valid rows for out-of-range lng, got {len(valid)}"
-    assert len(invalid) == 1, f"Expected 1 invalid row for out-of-range lng, got {len(invalid)}"
+    assert len(valid) == 0, (
+        f"Expected no valid rows for out-of-range lng, got {len(valid)}"
+    )
+    assert len(invalid) == 1, (
+        f"Expected 1 invalid row for out-of-range lng, got {len(invalid)}"
+    )
 
 
 def test_validate_rejects_non_numeric() -> None:
     rows: list[dict[str, object]] = [{"id": "a", "latitude": "bad", "longitude": 139.0}]
     valid, invalid = validate_coordinates(rows)
-    assert len(valid) == 0, f"Expected no valid rows for non-numeric lat, got {len(valid)}"
-    assert len(invalid) == 1, f"Expected 1 invalid row for non-numeric lat, got {len(invalid)}"
+    assert len(valid) == 0, (
+        f"Expected no valid rows for non-numeric lat, got {len(valid)}"
+    )
+    assert len(invalid) == 1, (
+        f"Expected 1 invalid row for non-numeric lat, got {len(invalid)}"
+    )
 
 
 def test_validate_accepts_int_coords() -> None:
     rows: list[dict[str, object]] = [{"id": "a", "latitude": 35, "longitude": 139}]
     valid, invalid = validate_coordinates(rows)
     assert len(valid) == 1, f"Expected 1 valid row for int coords, got {len(valid)}"
-    assert len(invalid) == 0, f"Expected no invalid rows for int coords, got {len(invalid)}"
+    assert len(invalid) == 0, (
+        f"Expected no invalid rows for int coords, got {len(invalid)}"
+    )
 
 
 def test_validate_empty_input() -> None:
@@ -113,7 +131,9 @@ def test_validate_mixed_rows() -> None:
     ]
     valid, invalid = validate_coordinates(rows)
     assert len(valid) == 1, f"Expected 1 valid row in mixed input, got {len(valid)}"
-    assert len(invalid) == 2, f"Expected 2 invalid rows in mixed input, got {len(invalid)}"
+    assert len(invalid) == 2, (
+        f"Expected 2 invalid rows in mixed input, got {len(invalid)}"
+    )
 
 
 # ── Clustering tests ─────────────────────────────────────────────────
